src/cdk/requirements.txt
aws-cdk.aws-apigateway
aws-cdk.aws-cloudtrail
aws-cdk.aws-cognito
aws-cdk.aws-ec2
aws-cdk.aws-elasticsearch
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-iam
aws-cdk.aws-kms
aws-cdk.aws-lambda
aws-cdk.aws-route53
aws-cdk.aws-route53-targets
aws-cdk.aws-s3
aws-cdk.aws-secretsmanager
aws-cdk.aws-sns
aws-cdk.aws-ssm
aws-cdk.aws-wafv2
aws-cdk.core
